refactor(server): extract cli entrypoint module
